@@ -1,53 +1,42 @@
-'use strict';
-
-var util = require('util');
-
-function CustomError(message, status_code) {
-    Error.call(this, message);
-    
-    if (Error.captureStackTrace) {
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+class CustomError extends Error {
+    constructor(message, code = 0) {
+        super(message);
+        this.code = code;
         Error.captureStackTrace(this, this.constructor);
-    } else {
-        this.stack = (new Error()).stack;
+        this.name = 'CustomError';
+        this.time = new Date();
     }
-    
-    this.message = message;
-    this.status_code = status_code;
-    this.time = new Date();
 }
-
-util.inherits(CustomError, Error);
-
-function Err(message, code) {
-    CustomError.call(this, message, 500);
-    
-    this.code = code;
-}
-util.inherits(Err, CustomError);
-
-function ValidationError(errors) {
-    CustomError.call(this, 'Validation Failed', 422);
-    
-    this.errors = errors;
-}
-util.inherits(ValidationError, CustomError);
-
+exports.CustomError = CustomError;
 /**
  *
- * @param status_code HTTP status code, such as 403, 404, 500, etc.
+ * @param statusCode HTTP status code, such as 403, 404, 500, etc.
  * @param message
  * @param code
- * @constructor
  */
-function HttpError(status_code, message, code) {
-    CustomError.call(this, message, status_code);
-    
-    this.code = code;
+class HttpError extends CustomError {
+    constructor(statusCode = 500, message, code) {
+        super(message, code);
+        this.statusCode = statusCode;
+        this.name = 'HttpError';
+    }
 }
-util.inherits(HttpError, CustomError);
-
-
-exports.CustomError = CustomError;
-exports.Err = Err;
-exports.ValidationError = ValidationError;
 exports.HttpError = HttpError;
+class ServerError extends HttpError {
+    constructor(message, code) {
+        super(500, message, code);
+        this.name = 'ServerError';
+    }
+}
+exports.ServerError = ServerError;
+class ValidationError extends HttpError {
+    constructor(errors, message = 'Validation Failed', statusCode = 422, code) {
+        super(statusCode, message, code);
+        this.errors = errors;
+        this.name = 'ValidationError';
+    }
+}
+exports.ValidationError = ValidationError;
+//# sourceMappingURL=errors.js.map
